Update file folder
